Microversion 2.71 - show server group
Add support microversion 2.71 which adds server group
information in the output of the following commands.

* nova show
* nova rebuild

The 'nova update' command does not output its result
when it is successful.
So there is no change for the command.

The patch for microversion 2.71 in the nova side is
I4a2a584df56ece7beb8b12c0ce9b0e6b30237120.

Change-Id: Id324486b5ef32615881085cd46772aa55c245ac6
Implements: blueprint show-server-group