Revert "Add irrelevant files in dsvm job"
This reverts commit 52ae954bd09bbb1536bc03b0be334d02e66fe93c.

The change ignored way too many things that should be run
through the functional dsvm job, such as changes to the
functional tests themselves, requirements, setup.cfg and
tox.ini changes. The bug was about not running the dsvm
functional job on docs changes and then went much further
and clearly not much thought was put into what was being
ignored and if it should have been, so let's just revert
and have a do-over.

Change-Id: Ie78badfeed7e51f1786eb5041010c7d17b02e412