Follow up for microversion 2.75
This is a follow-up for microversion 2.75 -
Multiple API cleanup changes.

The base class of the TestFlavorNovaClientV274 class is
the TenantTestBase class.
And the base class of the TenantTestBase class is
the ClientTestBase class.
It is not necessary to use the TenantTestBase class
as the base class. So specify the ClientTestBase class
directly.

TrivialFix
Change-Id: I2cb971f46ba697d9386ca61b7f51169f02b605ab