Clarify --migration-type migration value as cold migration
This is a follow up to a review discussion [1] where the
"migration" value for the --migration-type option is for
a cold migration. This change just updates docs and help
strings.

An alternative is changing "migration" as a valid value to
"cold-migration" (or accept both) and under the covers treat
"cold-migration" as "migration" for the actual API call.

[1] https://review.opendev.org/#/c/675117/3/novaclient/v2/shell.py@5430

Related to blueprint more-migration-list-filters

Change-Id: I6baa6af8731252e3c4976db06f0ca9cdfcb5e2f1