When creating a client, pass the default logger
When a client is created without a logger, novaclient creates a
default logger, but it's then not used, because it's not passed
to the factory function. Because of that, all novaclient calls
are getting logged as 'keystoneauth.session' instead of
'novaclient.v2.client' as they should.

Closes-bug: 1921388
Change-Id: I53caceb08667eb12e27016731868e8015dd10e34
(cherry picked from commit d4399d3d36c456c1001b96178ce9b3e7ec082460)